pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-mail-agent"
version = "0.1.0"
description = "AI-powered email agent with senior developer persona"
requires-python = ">=3.11"
dependencies = [
    # LLM Frameworks
    "langchain>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-community>=0.3",
    "langgraph>=0.2",
    "autogen-agentchat>=0.4",

    # Anthropic
    "anthropic>=0.40",

    # Gmail
    "google-api-python-client>=2.100",
    "google-auth-httplib2>=0.2",
    "google-auth-oauthlib>=1.2",

    # Vector DB
    "chromadb>=0.5",
    "sentence-transformers>=3.0",

    # Utilities
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "markdown>=3.5",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.7",
    "mypy>=1.13",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
